chore: refactor video provider with proper disposing of the resources, refactor story progress bar, add story progress hook
